+++ docs/changelog.md
@@ -2,11 +2,35 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format. 
 
+## [0.8.0] - 2019-12-06
+
+### Changed
+- `refgenie build` command arguments naming scheme: `--{input_name} <path>` to `--files {input_name}=<path>`
+- `-r`/`--requirements` in `refgenie build` command to `-q`/`--requirements`
+- recipe format: requirements (both assets and inputs) are lists of dicts rather that lists of strings
+- `refgenie list` displays current server subscriptions
+
+### Added
+- `refgenie id` command for asset digest retrieval
+- cross-namespace asset relationships support 
+- `--assets` argument in `refgenie build` command to provide parent assets, if required
+- `-r`/`--recipe` in `refgenie build` command argument to provide the recipe for the build
+- `subscribe` and `unsubscribe` subcommands to enable server list manipulation in the config file (`genome_servers` entry in the refgenie configuration file)
+- new recipes:
+    - `salmon_sa_index`
+    - `salmon_partial_sa_index`
+    - `suffixerator_index`
+    - `tallymer_index`
+
+### Removed
+- `-t`/`--tag` in `refgenie build`. Use more flexible `--assets` instead.
+
 ## [0.7.2] - 2019-11-06
 
 ### Added
 - `dbsnp` recipe
 - distribute the license file with the package
+
 
 ## [0.7.1] - 2019-10-29